Move sim logic from AddMoteDialog
Have the dialog return the settings
and put the simulation updates inside
GUI. The logic does not really
belong in GUI either, but it is
a better place than in the dialog.
